Quick changes
ghosts are now stored in a vector to allow easier additions and
removals. Added functionality for ghosts to be added and deleted when a
power dot is eating for demo purposes. Found a bug that occurs
sporadically, there seems to be an issue with the amount of directions
available for the ghosts. Haven't found the exact issue because it
doesn't occur every run.