Pacman Changes Colors and Music Files
Made changes so that Pacman will change color based on emotional state.
I was unable to test if this works in real time due to an issue with my
environment.jar file. I also added the music files to the location they
needed to be in for my project to actually play the sounds. This may be
different for other users but it seemed like adding these wouldn't hurt
anything.