Added Log File and Emojis
A basic log file is now produced as the emotional states are read.
Emojis are also displayed on screen to show the current user state.
These images will be altered later to remove white backgrounds but will
do for now. A flag is also being messed with but should be implemented
in a later release. Bugs were removed regarding Pacman changing colors.
Additional functions were added for convenience.